Keep the default encode path baseline-identical with Marshaler support
Benchmarking the Marshaler-interface commit (#1084) against v2 on
linux/amd64 (interleaved rounds, n=12, benchstat) showed statistically
significant sec/op regressions on the default (interface-disabled)
Marshal path — up to +10% on micro-benchmarks — with allocations
unchanged. Bisecting the diff piece by piece isolated the dominant
cause: widening the per-type typeEncProps cache struct from 2 to 3
bytes. Its cached p.(typeEncProps) assertion and by-value return run
two to three times per encoded value, and the extra field changes the
copy size and register shape of every lookup; that lone change
reproduced the regression with zero encoder-code modifications.

- typeEncProps and encPropsForType are restored to their exact baseline
  form. The unstable.Marshaler classification lives in its own
  sync.Map (typeMarshalerCache/marshalerPropsForType), consulted only
  on EnableMarshalerInterface paths.
- encodeTable, encodeKeyValue and encodeArrayTable are restored to
  their exact baseline source. The Marshaler-aware variants of the
  table passes live in encodeTableWithMarshalers and
  encodeArrayTableWithMarshalers, selected by one branch per table
  instead of one per entry. Marshaler values on a `key =` line flow
  through appendValue's gated guard, which returns without rejoining
  the fall-through path and re-classifies the fetched bytes (keeping
  the emit-time guard), so appendMarshalerInlineValue is gone.
- All Marshaler-interface machinery moves to a new marshaler_raw.go so
  the feature's cold code does not sit between the default path's hot
  functions.

After this change the interface-off encoder differs from baseline only
by four single-branch dispatches; typeEncProps, encPropsForType and the
hot encoder functions are byte-identical to the pre-#1084 source.
allocs/op are identical across the whole ./benchmark suite. Remaining
µs-scale sec/op deltas are within this host's address-alignment noise
floor, demonstrated by dead-code A/A controls and -randlayout runs.
